+++ exportplan/serializers.py
@@ -1,8 +1,7 @@
 from django.contrib.postgres.fields import JSONField
-from django.db import transaction
 from rest_framework import serializers
 
-from exportplan import models
+from exportplan import helpers, models
 
 
 class CompanyObjectivesSerializer(serializers.ModelSerializer):
@@ -118,6 +117,36 @@ class ExportPlanDownloadSerializer(serializers.ModelSerializer):
         }
 
 
+class ExportPlanListSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = models.CompanyExportPlan
+        fields = (
+            'pk',
+            'name',
+            'created',
+            'ui_progress',
+            'export_countries',
+            'export_commodity_codes',
+        )
+
+
+class ExportPlanCreateSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = models.CompanyExportPlan
+        fields = (
+            'pk',
+            'sso_id',
+            'export_countries',
+            'export_commodity_codes',
+        )
+
+    def create(self, validated_data):
+        validated_data['name'] = helpers.get_unique_exportplan_name(validated_data)
+        new_plan = models.CompanyExportPlan(**validated_data)
+        new_plan.save()
+        return new_plan
+
+
 class CompanyExportPlanSerializer(serializers.ModelSerializer):
     company_objectives = CompanyObjectivesSerializer(many=True, required=False, read_only=False)
     route_to_markets = RouteToMarketsSerializer(many=True, required=False, read_only=False)
@@ -129,6 +158,7 @@ class CompanyExportPlanSerializer(serializers.ModelSerializer):
     class Meta:
         model = models.CompanyExportPlan
         fields = (
+            'name',
             'company',
             'sso_id',
             'ui_options',
@@ -167,13 +197,6 @@ class CompanyExportPlanSerializer(serializers.ModelSerializer):
             serializer.is_valid(raise_exception=True)
         return value
 
-    def create(self, validated_data):
-
-        objectives = validated_data.pop('company_objectives', {})
-        instance = super().create(validated_data)
-        self.recreate_objectives(instance, objectives)
-        return instance
-
     def update(self, instance, validated_data):
 
         # This will allow partial updating to json fields during a patch update. Json fields generally represent
@@ -200,12 +223,3 @@ class CompanyExportPlanSerializer(serializers.ModelSerializer):
                 validated_data[field_name] = field_value
         super().update(instance, validated_data)
         return instance
-
-    @transaction.atomic
-    def recreate_objectives(self, instance, objectives):
-        instance.company_objectives.all().delete()
-        data_collection = []
-        for objective in objectives:
-            data = {**objective, 'companyexportplan': instance}
-            data_collection.append(models.CompanyObjectives(**data))
-        models.CompanyObjectives.objects.bulk_create(data_collection)