+++ CHANGELOG.md
@@ -3,11 +3,23 @@
 ## Pre release
 ### Bugs fixed
 ### Enhancements
+- GP2-3343 - Making unsubscribe token based
+- GP2-3275 - Export plan delete
+- GP2-3171 - Add export plan PK to serializer
+- GP2-1319 - relabelling/adding data services models's attributes
+- GP2-3173 - multi-Export-plan-list-detail
+- GP2-3351 - Add permission class to export-plan api can only be changed/accessed by owner 
+- GP2-3179 - Spike management command: report_export_plan
 
+
+## [1.14.0](https://github.com/uktrade/directory-api/releases/tag/1.14.0)
+[Full Changelog](https://github.com/uktrade/directory-api/compare/1.13.0...1.14.0)
 - GP2-2867 - dockerise d-api
 - NO TICKET - Firebreak week allow SIC Codes to be stored and updated
 - GP2-1319 - Relabelling data services models
 - GP2-2896 - Split company objectives start/end date fields
+- GP2-3179 - Spike management command: report_export_plan
+- [HOTFIX] - GP2-3446 change london ita search
 
 ## [1.13.0](https://github.com/uktrade/directory-api/releases/tag/1.12.0)
 [Full Changelog](https://github.com/uktrade/directory-api/compare/1.12.0...1.13.0)